Adds mechanics to extend document loaders and class with plugins
The docs need to be amended in this regard.
