+++ delb/plugins/__init__.py
@@ -1,0 +1,25 @@
+# Copyright (C) 2019  Frank Sachsenheim
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import pluggy  # type: ignore
+
+from delb.plugins import specs
+from delb.plugins.contrib import core_loaders
+
+plugin_manager = pluggy.PluginManager("delb")
+plugin_manager.add_hookspecs(specs)
+plugin_manager.load_setuptools_entrypoints("delb")
+plugin_manager.register(core_loaders)
+plugin_manager.check_pending()
